telegram-bot-docker/requirements.txt
pyTelegramBotAPI==4.14.0
google-generativeai
aiohttp
python-telegram-bot==21.1.1
